# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESTFull API Lumen",
        "description": "RESTFull API SwaggerOpenApi description.",
        "contact": {
            "email": "itv.develop@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/{id}/companies": {
            "get": {
                "tags": [
                    "User company"
                ],
                "summary": "Get user company",
                "operationId": "UserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Api token can be obtained by login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            },
            "post": {
                "tags": [
                    "User company"
                ],
                "summary": "Add company for user",
                "operationId": "UserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Api token can be obtained by login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title company",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone company",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description company",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            }
        },
        "/api/recover-password": {
            "post": {
                "tags": [
                    "Reset password user"
                ],
                "summary": "Send reset link by email",
                "operationId": "App\\Http\\Controllers\\SendResetPasswordController::__construct",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            }
        },
        "/api/sign-in": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login user into system",
                "operationId": "App\\Http\\Controllers\\UserController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registration user into system",
                "operationId": "App\\Http\\Controllers\\UserController::register",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name for new user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name for new user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email for new user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password for new user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone for new user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "RESTFull API",
            "description": "API Endpoints to share the company's information for the logged users"
        }
    ]
}